+++ src/Kairo/router/init/AddonInitializeActivator.ts
