+++ src/Kairo/router/init/AddonInitializeReceive.ts
