+++ src/Kairo/router/init/AddonInitializeRegister.ts
