+++ src/Kairo/router/init/AddonInitializeRequest.ts
