+++ src/Kairo/router/init/AddonInitializeResponse.ts
